Specify the required webpack version
The interface doesn't currently work with the latest webpack version.
However, it's this version which is being installed when the docker
image is built or the dev setup instructions are followed. This causes
various errors and prevents the CSS loading correctly (c.f. #55).

For now, update the instructions and docker image so anyone following
them will install a version that the UI works with. This means people
don't need to waste time trying to get their setup working.