Update App Details page to use V2 API
The App Details page was not working because it was trying to use
functionality from the old V1 API (for example fn routes). This commit
changes the App Details page to use V2's fns API endpoint and removes
the routes logic. Users can now add/edit/remove/run functions for a
particular app.

This contributes to: #52, fnproject/fn#1347, and fnproject/fn#1258.