Add busy function chart to App Details page
The App Details page used to display charts that showed metrics on the
functions that are part of the App. However, the fn metrics API changed
and no longer included this information so it was removed. In
fnproject/fn#1365 some very similar metrics were added to the metrics
API so add the chart showing the currently running functions back and
update it to use this metric. Furthermore, refactor controllers/stats.js
to make it easier to parse additional stats in the future.

This contributes to: #52 and fnproject/fn#1258.