Use Fn names rather than image names in charts
I'm not convinced the App Details graphs are going to scale when there
are lots of functions. By using the function name rather than breaking
it down into its own individual images it should improve performance a
little bit as there is less to plot (although maybe not enough).

Furthermore, it seems unlikely old images will be used so it's wasted
information.

Lastly, it might be possible for functions to share images which would
make it difficult to differentiate between these.