Add support for custom statistic tags to the graphs
In fnproject/fn#1438 Reed made the tags shown in Fn's metrics API
configurable. This means in some custom builds there may not be an
appID, fnID, or imageName attribute associated with the fn_container
stats. This issue was brought to light in #58.

Update the code so that when the Fn server doesn't provide the required
information the interface handles it gracefully, showing a message with
possible reasons as to why the stats may not be showing.