Fix NaN bug on graphs
Sometimes the Fn Server doesn't return data which a graph uses for
example 'fn_container_paused' isn't shown until a function enters the
state. This means the Paused graph shows "NaN" and doesn't plot a line.
Update the graphs so that when there is no metric, it defaults to 0
rather than undefined.