Move Fn/App config to its own Vue component
The Fn/App config was duplicated on both the App and Fn form. By moving
it to its own component it means we only have to maintain it in one
place. For example, I recently fixed the deletion of config lines.
Having one component to edit means I only need to apply the fixes to one
place.

As a consequence of this, it also means it fixes the deletion of config
points for Apps - prior to this change deleting the config row wouldn't
have any effect and you'd need to set the config value to an empty
string to delete it.