Fix bug when editing config keys (#66)
There is a bug which means that when a config key is edited a new config
is added but the old config isn't deleted. Make the config keys readonly
so that it forces the user to delete the old config value and add a new
one rather than editing an existing config which will cause issues.

Fixes #66.