Use config-logger rather than console.log()
Using console.log was causing eslint to complain. Using console.log is
against best practise.

Although in these particular cases the logs are going to the Node
server rather than to a client's console. The rest of the code uses
config-logger so update the console.log() calls to also use this to be
consistent. This also provides a cleaner eslint output meaning future
warnings/errors are less likely to be ignored and means logging is more
configurable.