Only update Docker in CircleCI on deployments
We don't use Docker until we deploy the app. This stage takes a minute
to run so only run it if we're on the master branch (and therefore
deploying the app). This means we will get faster feedback e.g. for
failed tests and reduces the amount of CircleCI compute time we use