Don't use dev packages in production
The package.json file is split between dev and production packages
meaning we can configure which packages are installed in the production
image. Update the docker file so that it makes use of this
functionality. This reduces the docker image's size and also reduces the
attack space from vulnerable packages