Use seperate modules for HTML element details for integration tests
This commit adds a new Element Details class which can store information
on how to locate an HTML element on the page. It also separates the
Element Details from the code into its own seperate modules so we can
keep all of the HTML element details together in one place. It also means
we only need to change one place if the HTML element's details change. For
example, before, if the input field for a Function's image had its ID changed,
we'd need to update it in multiple places.